refactor(combo): extract target sorters to combo/targetSorters.ts
